Cleaned up the project by moving code into seperate files and deleted unnecessary files
