+++ yolo.py
@@ -1,0 +1,106 @@
+import numpy as np
+import cv2
+
+# Remove the bounding boxes with low confidence using non-maxima suppression
+# image: image detection performed on
+# results: output from YOLO CNN network
+# threshold_confidence: threshold on keeping detection
+# threshold_nms: threshold used in non maximum suppression
+
+def initialise(class_file, config_file, weights_file):
+    
+    # init YOLO CNN object detection model
+
+    confThreshold = 0.5  # Confidence threshold
+    nmsThreshold = 0.4   # Non-maximum suppression threshold
+    inpWidth = 416       # Width of network's input image
+    inpHeight = 416      # Height of network's input image
+
+    # Load names of classes from file
+
+    classesFile = class_file
+    classes = None
+    with open(classesFile, 'rt') as f:
+        classes = f.read().rstrip('\n').split('\n')
+
+    # load configuration and weight files for the model and load the network using them
+
+    net = cv2.dnn.readNetFromDarknet(config_file, weights_file)
+    output_layer_names = getOutputsNames(net)
+
+    # defaults DNN_BACKEND_INFERENCE_ENGINE if Intel Inference Engine lib available or DNN_BACKEND_OPENCV otherwise
+    net.setPreferableBackend(cv2.dnn.DNN_BACKEND_DEFAULT)
+
+    # change to cv2.dnn.DNN_TARGET_CPU (slower) if this causes issues (should fail gracefully if OpenCL not available)
+    net.setPreferableTarget(cv2.dnn.DNN_TARGET_OPENCL)
+
+    return inpWidth, inpHeight, classes, net, output_layer_names
+
+
+def postprocess(image, results, threshold_confidence, threshold_nms):
+    frameHeight = image.shape[0]
+    frameWidth = image.shape[1]
+
+    # Scan through all the bounding boxes output from the network and..
+    # 1. keep only the ones with high confidence scores.
+    # 2. assign the box class label as the class with the highest score.
+    # 3. construct a list of bounding boxes, class labels and confidence scores
+
+    classIds = []
+    confidences = []
+    boxes = []
+    for result in results:
+        for detection in result:
+            scores = detection[5:]
+            classId = np.argmax(scores)
+            confidence = scores[classId]
+            if confidence > threshold_confidence:
+                center_x = int(detection[0] * frameWidth)
+                center_y = int(detection[1] * frameHeight)
+                width = int(detection[2] * frameWidth)
+                height = int(detection[3] * frameHeight)
+                left = int(center_x - width / 2)
+                top = int(center_y - height / 2)
+                classIds.append(classId)
+                confidences.append(float(confidence))
+                boxes.append([left, top, width, height])
+
+    # Perform non maximum suppression to eliminate redundant overlapping boxes with
+    # lower confidences
+    classIds_nms = []
+    confidences_nms = []
+    boxes_nms = []
+
+    indices = cv2.dnn.NMSBoxes(boxes, confidences, threshold_confidence, threshold_nms)
+    for i in indices:
+        i = i[0]
+        classIds_nms.append(classIds[i])
+        confidences_nms.append(confidences[i])
+        boxes_nms.append(boxes[i])
+
+    # return post processed lists of classIds, confidences and bounding boxes
+    return (classIds_nms, confidences_nms, boxes_nms)
+
+
+def getOutputsNames(net):
+    # Get the names of all the layers in the network
+    layersNames = net.getLayerNames()
+    # Get the names of the output layers, i.e. the layers with unconnected outputs
+    return [layersNames[i[0] - 1] for i in net.getUnconnectedOutLayers()]
+
+
+def create_and_remove(cropped_img, clahe_img, inpWidth, inpHeight, net, output_layer_names):
+
+     # create a 4D tensor (OpenCV 'blob') from image frame (pixels scaled 0->1, image resized)
+    tensor = cv2.dnn.blobFromImage(cropped_img, 1/255, (inpWidth, inpHeight), [0,0,0], 1, crop=False)
+
+    # set the input to the CNN network
+    net.setInput(tensor)
+
+    # runs forward inference to get output of the final output layers
+    results = net.forward(output_layer_names)
+
+    # remove the bounding boxes with low confidence
+    classIDs, confidences, boxes = postprocess(clahe_img, results, 0.5, 0.4)
+
+    return classIDs, boxes
